requirements.txt
requests
python-telegram-bot==12.6.1 
google-api-python-client>=1.7.11,<1.7.20
google-auth-httplib2>=0.0.3,<0.1.0
google-auth-oauthlib>=0.4.1,<0.10.0
python-dotenv>=0.10
telegraph